+++ Core/Constants.lua
@@ -5,7 +5,7 @@ ns.CONST = {
 	ADDON_NAME       = "CooldownMaster",
 	ADDON_DISPLAY    = "Cooldown Master",
 	ADDON_SHORT      = "CDM",
-	VERSION          = "1.12.0",
+	VERSION          = "1.13.0",
 
 	SLASH_COMMANDS   = { "cdmaster", "cooldownmaster", "cm" },